change the return value of get_fs_type from int to long.
The field fs_type field of fd_parms is already long
Additionally the glibc man page says "Using unsigned int for such variables suffices on most systems." referring the minimum width of variables holding fs_type
There are already fs_types who sets the highest bit (CIFS, BTRFS HPFS...) which interfere with the error check in dump_one_reg_file_cond (fs_type < 0) through the cast into signed int and after this into long

Fixes error while dumping programs on a btrfs partition

+++ criu/cr-dump.c
@@ -229,7 +229,7 @@ static int fill_fd_params_special(int fd, struct fd_parms *p)
 	return 0;
 }
 
-static int get_fs_type(int lfd)
+static long get_fs_type(int lfd)
 {
 	struct statfs fst;
